integration: move watch tests to v3_watch_test.go
